added feature
ha_to_m2_converter added in input

+++ p1-a310a.c
@@ -20,8 +20,8 @@ struct area {
   int pH;
   int nutrient;
   int moistness;
-  double totalArea;
-  double mfoArea;
+  int totalArea;
+  int mfoArea;
 };
 
 enum mfoTypes {
